[Refactor] 이벤트관련 네이밍, 위치 수정 (#43)
- 이벤트발행을 위한 클래스는 application 쪽에 위치
- 주문접수 이벤트에 해당하는 클래스는 domain entity
