+++ application-client/src/main/kotlin/org/fastcampus/applicationclient/order/service/event/OrderNotificationEvent.kt
@@ -1,0 +1,7 @@
+package org.fastcampus.applicationclient.order.service.event
+
+import org.fastcampus.order.entity.Order
+
+data class OrderNotificationEvent(
+    val order: Order,
+)
